[AIRFLOW-5895] Move HDFS stuff from tests/core.py
