+++ tests/hooks/test_webhdfs_hook.py
@@ -95,3 +95,11 @@ class TestWebHDFSHook(unittest.TestCase):
             overwrite=True,
             n_threads=1
         )
+
+    def test_simple_init(self):
+        c = WebHDFSHook()
+        self.assertIsNone(c.proxy_user)
+
+    def test_init_proxy_user(self):
+        c = WebHDFSHook(proxy_user='someone')
+        self.assertEqual('someone', c.proxy_user)
