+++ tests/sensors/test_hdfs_sensor.py
@@ -19,10 +19,11 @@
 import unittest
 from datetime import timedelta
 
+from tests.test_utils.hdfs_utils import FakeHDFSHook
+
 from airflow.exceptions import AirflowSensorTimeout
 from airflow.sensors.hdfs_sensor import HdfsSensor
 from airflow.utils.timezone import datetime
-from tests.core import FakeHDFSHook
 
 DEFAULT_DATE = datetime(2015, 1, 1)
 TEST_DAG_ID = 'unit_test_dag'
